fbdev: msm: fix compilation issue in 32-bit mode
Fix typecasting issue specific to 32-bit mode.

+++ drivers/video/fbdev/msm/mdss_fb.h
@@ -307,7 +307,7 @@ struct msm_fb_data_type {
 	u32 calib_mode;
 	u32 calib_mode_bl;
 	u32 ad_bl_level;
-	u32 bl_level;
+	u64 bl_level;
 	u32 bl_scale;
 	u32 bl_min_lvl;
 	u32 unset_bl_level;
